pyproject.toml
[tool.poetry]
name = "tomatempo"
version = "0.1.0"
description = "An Pomodoro based application for time management."
authors = ["André Carvalho <alac1984@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlmodel = "^0.0.24"
alembic = "^1.16.4"
typer = "^0.16.0"
platformdirs = "^4.3.8"
pydantic-settings = "^2.10.1"
pyyaml = "^6.0.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pynvim = "^0.5.2"
mypy = "^1.17.1"
freezegun = "^1.5.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["src"]
markers = [
    "smoke: test smoke imports",
    "slow: mark test as slow."
]

[tool.mypy]
mypy_path = "src"
namespace_packages = true
explicit_package_bases = true

[tool.ruff]
# Ajusta à sua versão (py311, py312…)
target-version = "py311"
line-length = 100
# arquivos/pastas a ignorar
exclude = ["build", "dist", ".venv", ".mypy_cache", ".ruff_cache"]

[tool.ruff.lint]
select = [
  "E",   # pycodestyle
  "F",   # pyflakes
  "W",   # warnings
  "I",   # isort (ordena imports)
  "UP",  # pyupgrade
  "B",   # bugbear
  "C4",  # list/set/dict comprehensions
  "PT",  # convenções pytest
]
ignore = [
  "E501",  # comprimento de linha (o formatter já resolve)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
